+++ pyvirtualdisplay/examples/screenshot1.py
@@ -1,0 +1,10 @@
+from easyprocess import EasyProcess
+from pyvirtualdisplay.smartdisplay import SmartDisplay
+
+disp = SmartDisplay(visible=0, bgcolor='black').start()
+xmessage = EasyProcess('xmessage screenshot').start()
+img = disp.waitgrab()
+xmessage.stop()
+disp.stop()
+img.show()
+