+++ CHANGES.rst
@@ -1,3 +1,11 @@
+Release 0.0.2 (2/23/2011)
+================================
+* more doc
+* more examples
+* Display is class
+* new: SmartDisplay class for screenshots
+
+
 Release 0.0.1 (2/11/2011)
 ================================